added

+++ casualbear_api/app/models/event.js
@@ -2,7 +2,7 @@ module.exports = (sequelize, DataTypes) => {
   const Event = sequelize.define("Event", {
     name: DataTypes.STRING,
     description: DataTypes.STRING,
-    selectedColor: DataTypes.INTEGER,
+    selectedColor: DataTypes.BIGINT,
     rawUrl: DataTypes.STRING,
   });
